[UPDATE] 自動収納切り替えボタンを別オブジェクトに分離する
カテゴリ内でもこのボタンが使われているらしい
